Minor tweaks to handling of incorrectly formatted .mpc lines and program continuation when errors are encountered.

+++ src/ossos/core/tools/postage_stamp_builder.py
@@ -19,34 +19,48 @@ from astropy.units import Quantity
 from astropy.io import fits
 from ossos import mpc
 from ossos import storage
+from planning.plotting import parameters
 
 
 def cutout(obj, obj_dir, radius):
+    print len([n for n in obj.mpc_observations if not n.null_observation])
     for obs in obj.mpc_observations:
+        print('starting analysis of {}'.format(str(obs)))
         if obs.null_observation:
+            print('skipping')
             continue
-        expnum = obs.comment.frame.split('p')[0]  # only want calibrated images
-        if not expnum.isdigit():
-            logging.error('expnum {} parsed from comment line invalid. Check comment parsing.\n{}'.format(expnum, str(
-                obs.comment)))
-            continue
-        uri = storage.get_uri(expnum)
-        sky_coord = obs.coordinate  # Using the WCS rather than the X/Y (X/Y can be unreliable over the whole survey)
-        logging.info('Trying {} on {} on {}...\n'.format(obj.provisional_name, obs.date, expnum))
-        hdulist = storage.ra_dec_cutout(uri, sky_coord, radius)
-        # if not 'ASTLEVEL' in hdulist[1].header:   # FIXME: activate once all headers are retro-fitted with ASTLEVEL
-        #     logging.info('Cutout invalid for use. Skipping inclusion.\n')
-        #     continue
-        postage_stamp_filename = "{}_{:11.5f}_{:09.5f}_{:+09.5f}.fits".format(obj.provisional_name,
-                                                                              obs.date.mjd,
-                                                                              obs.coordinate.ra.degree,
-                                                                              obs.coordinate.dec.degree)
-        logging.info("{}".format(postage_stamp_filename))
+        if obs.date > parameters.SURVEY_START:  # can't make postage stamps of earlier linkages
+            # can't parse for an obs.comment's exposure number if no obs.comment exists
+            try:
+                expnum = obs.comment.frame.split('p')[0].strip(' ')  # only want calibrated images
+            except AttributeError, e:
+                print('No comment in this MPC line!')
+                continue
+            if not expnum.isdigit():
+                print('expnum {} parsed from comment line invalid. Check comment parsing.\n{}'.format(expnum, str(
+                    obs.comment)))
+                continue
+            uri = storage.get_uri(expnum)
+            sky_coord = obs.coordinate  # Using the WCS rather than the X/Y (X/Y can be unreliable over the whole survey)
+            print('Trying {} on {} on {}...'.format(obj.provisional_name, obs.date, expnum))
+            try:
+                hdulist = storage.ra_dec_cutout(uri, sky_coord, radius)
+                # if not 'ASTLEVEL' in hdulist[1].header:   # FIXME: activate once all headers are retro-fitted with ASTLEVEL
+                #     logging.info('Cutout invalid for use. Skipping inclusion.\n')
+                #     continue
+                postage_stamp_filename = "{}_{:11.5f}_{:09.5f}_{:+09.5f}.fits".format(obj.provisional_name,
+                                                                                      obs.date.mjd,
+                                                                                      obs.coordinate.ra.degree,
+                                                                                      obs.coordinate.dec.degree)
+                print("{}".format(postage_stamp_filename))
 
-        with open(postage_stamp_filename, 'w') as tmp_file:
-            hdulist.writeto(tmp_file, clobber=True)
-            storage.copy(postage_stamp_filename, obj_dir + "/" + postage_stamp_filename)
-        os.unlink(postage_stamp_filename)  # easier not to have them hanging around
+                with open(postage_stamp_filename, 'w') as tmp_file:
+                    hdulist.writeto(tmp_file, clobber=True)
+                    storage.copy(postage_stamp_filename, obj_dir + "/" + postage_stamp_filename)
+                os.unlink(postage_stamp_filename)  # easier not to have them hanging around
+            except OSError, e:  # occasionally the node is not found: report and move on for later cleanup
+                print e
+                continue
 
 
 def main():
@@ -64,7 +78,7 @@ def main():
                              "holding the .ast files of astrometry/photometry measurements.")
     parser.add_argument("--blocks", "-b",
                         action="store",
-                        default=["o3o"],
+                        default=["o4h"],
                         choices=["o3e", "o3o", "Col3N", 'o3l', 'o4h'],
                         help="Prefixes of object designations to include.")
     parser.add_argument("--radius", '-r',
@@ -93,6 +107,7 @@ def main():
                     storage.mkdir(obj_dir)
                 # assert storage.exists(obj_dir, force=True)
                 print('{} beginning...\n'.format(obj.provisional_name))
+                # if int(obj.provisional_name[3:]) == 49:
                 assert isinstance(args.radius, Quantity)
                 cutout(obj, obj_dir, args.radius)
                 print('{} complete.\n'.format(obj.provisional_name))
